requirements.txt
datasets==3.2.0
flax==0.10.2
jax==0.5.0
jaxopt==0.8.3
libsvmdata==0.4.1
matplotlib==3.10.0
numpy==2.2.2
optax==0.2.4
pandas==2.2.3
scikit_learn==1.6.1
scipy==1.15.1
tensorflow_cpu==2.18.0
tensorflow_datasets==4.9.7
tqdm==4.67.1
wandb==0.19.5
